requirements.txt
basicsr==1.4.2
kornia==0.5.1
face-alignment==1.3.4
ninja==1.10.2.3
einops==0.4.1
facexlib==0.2.5
librosa==0.9.2
dlib==19.24.0
numpy==1.22.0
moviepy==1.0.3
python-dotenv==1.0.1
ffmpeg-python==0.2.0
boto3==1.34.161
flask==3.0.3
gevent==24.2.1
grpcio-tools==1.65.5
grpcio==1.65.5